🚚 Move add_labels and get_labels from FeatureManager to Data
