+++ lamindb/_file.py
@@ -14,11 +14,9 @@ from lamindb_setup._init_instance import register_storage
 from lamindb_setup.dev import StorageSettings
 from lamindb_setup.dev._docs import doc_args
 from lnschema_core import Feature, FeatureSet, File, Run, Storage, ids
-from lnschema_core.models import Data
 from lnschema_core.types import AnnDataLike, DataLike, PathLike
 
 from lamindb._context import context
-from lamindb.dev import FeatureManager
 from lamindb.dev._settings import settings
 from lamindb.dev.hashing import b16_to_b64, hash_file
 from lamindb.dev.storage import (
@@ -962,15 +960,6 @@ def inherit_relations(self, file: File, fields: Optional[List[str]] = None):
         )
 
 
-@property  # type: ignore
-@doc_args(Data.features.__doc__)
-def features(self) -> "FeatureManager":
-    """{}"""
-    from lamindb._feature_manager import FeatureManager
-
-    return FeatureManager(self)
-
-
 METHOD_NAMES = [
     "__init__",
     "from_anndata",
@@ -1004,5 +993,3 @@ File._save_skip_storage = _save_skip_storage
 # TODO: move these to METHOD_NAMES
 setattr(File, "view_lineage", view_lineage)
 setattr(File, "inherit_relations", inherit_relations)
-# property signature is not tested:
-setattr(Data, "features", features)
